Cleaned up naming convention.
