+++ client/src/components/Attacks.js
@@ -2,31 +2,31 @@ import React from 'react'
 
 
 
-export default class Child11 extends React.Component {
-    state = {spells: [""]}
+export default class Attacks extends React.Component {
+    state = {attacks: [""]}
     
     handleInputChange(event) {
-        const state = this.state.spells
+        const state = this.state.attacks
         state.push(event.target.value)
-        this.setState({spells: state})
+        this.setState({attacks: state})
     }
 
     handleChange(event, index) {
-        const state = this.state.spells.slice()
+        const state = this.state.attacks.slice()
         state[index] = event.target.value
-        this.setState({spells: state})
+        this.setState({attacks: state})
     }
 
     handleAddFields() {
-        const state = this.state.spells
+        const state = this.state.attacks
         state.push("")
-        this.setState({spells: state})
+        this.setState({attacks: state})
       };
 
       handleRemoveFields(index) {
-        const state = this.state.spells
+        const state = this.state.attacks
         state.splice(index, 1)
-        this.setState({spells: state})
+        this.setState({attacks: state})
       };
 
       
@@ -35,16 +35,16 @@ export default class Child11 extends React.Component {
         return( 
             <form>
                 <div className="form-row">
-                    {this.state.spells.map((spells, index) => (
+                    {this.state.attacks.map((attacks, index) => (
                     <div className="form-group col-sm-6">
-                        <label htmlFor="spells">Spells</label>
+                        <label htmlFor="attacks">Attacks</label>
                         <input
                         type="text"
                         className="form-control"
-                        id="spells"
-                        name="spells"
+                        id="attacks"
+                        name="attacks"
                         onChange = {event => this.handleChange(event, index)}
-                        value={spells}
+                        value={attacks}
                         />
                     <button className="btn btn-link" type="button" onClick={() => this.handleRemoveFields(index)}> Delete </button>
                     </div>
