+++ client/src/components/Equipment.js
@@ -2,31 +2,31 @@ import React from 'react'
 
 
 
-export default class Child9 extends React.Component {
-    state = {languages: [""]}
+export default class Equipment extends React.Component {
+    state = {equipment: [""]}
     
     handleInputChange(event) {
-        const state = this.state.languages
+        const state = this.state.equipment
         state.push(event.target.value)
-        this.setState({languages: state})
+        this.setState({equipment: state})
     }
 
     handleChange(event, index) {
-        const state = this.state.languages.slice()
+        const state = this.state.equipment.slice()
         state[index] = event.target.value
-        this.setState({languages: state})
+        this.setState({equipment: state})
     }
 
     handleAddFields() {
-        const state = this.state.languages
+        const state = this.state.equipment
         state.push("")
-        this.setState({languages: state})
+        this.setState({equipment: state})
       };
 
       handleRemoveFields(index) {
-        const state = this.state.languages
+        const state = this.state.equipment
         state.splice(index, 1)
-        this.setState({languages: state})
+        this.setState({equipment: state})
       };
 
       
@@ -35,16 +35,16 @@ export default class Child9 extends React.Component {
         return( 
             <form>
                 <div className="form-row">
-                    {this.state.languages.map((languages, index) => (
+                    {this.state.equipment.map((equipment, index) => (
                     <div className="form-group col-sm-6">
-                        <label htmlFor="languages">Languages</label>
+                        <label htmlFor="equipment">Equipment</label>
                         <input
                         type="text"
                         className="form-control"
-                        id="languages"
-                        name="laguages"
+                        id="equipment"
+                        name="equipment"
                         onChange = {event => this.handleChange(event, index)}
-                        value={languages}
+                        value={equipment}
                         />
                     <button className="btn btn-link" type="button" onClick={() => this.handleRemoveFields(index)}> Delete </button>
                     </div>
