+++ client/src/components/Spells.js
@@ -2,31 +2,31 @@ import React from 'react'
 
 
 
-export default class Child7 extends React.Component {
-    state = {equipment: [""]}
+export default class Spells extends React.Component {
+    state = {spells: [""]}
     
     handleInputChange(event) {
-        const state = this.state.equipment
+        const state = this.state.spells
         state.push(event.target.value)
-        this.setState({equipment: state})
+        this.setState({spells: state})
     }
 
     handleChange(event, index) {
-        const state = this.state.equipment.slice()
+        const state = this.state.spells.slice()
         state[index] = event.target.value
-        this.setState({equipment: state})
+        this.setState({spells: state})
     }
 
     handleAddFields() {
-        const state = this.state.equipment
+        const state = this.state.spells
         state.push("")
-        this.setState({equipment: state})
+        this.setState({spells: state})
       };
 
       handleRemoveFields(index) {
-        const state = this.state.equipment
+        const state = this.state.spells
         state.splice(index, 1)
-        this.setState({equipment: state})
+        this.setState({spells: state})
       };
 
       
@@ -35,16 +35,16 @@ export default class Child7 extends React.Component {
         return( 
             <form>
                 <div className="form-row">
-                    {this.state.equipment.map((equipment, index) => (
+                    {this.state.spells.map((spells, index) => (
                     <div className="form-group col-sm-6">
-                        <label htmlFor="equipment">Equipment</label>
+                        <label htmlFor="spells">Spells</label>
                         <input
                         type="text"
                         className="form-control"
-                        id="equipment"
-                        name="equipment"
+                        id="spells"
+                        name="spells"
                         onChange = {event => this.handleChange(event, index)}
-                        value={equipment}
+                        value={spells}
                         />
                     <button className="btn btn-link" type="button" onClick={() => this.handleRemoveFields(index)}> Delete </button>
                     </div>
